move styled button
