+++ src/app/(Authentication)/Welcome/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Text, View } from 'react-native';
-import StyledButton from '../../../Components/StyledButton';
+import StyledButton from '../../../Components/StyledButton/StyledButton';
 import styles from './styles';
 
 function WelcomeScreen() {
